requirements.txt
Open AI
Langchain
python-dotenv
